+++ Posh-ACME/DnsPlugins/Azure.ps1
@@ -282,9 +282,9 @@ function ConvertFrom-AccessToken {
     }
 
     # make sure the token hasn't expired
-    $expiresDT = [DateTime](Get-Date '1/1/1970').AddSeconds($claims.exp)
-    if ($expiresDT -lt [DateTime]::UtcNow) {
-        throw "The provided access token has expired as of $("{0:yyyy-MM-ddTHH:mm:ssZ}" -f $expiresDT)"
+    $expires = [DateTimeOffset]::FromUnixTimeSeconds($claims.exp)
+    if ((Get-DateTimeOffsetNow) -gt $expires) {
+        throw "The provided access token has expired as of $($expires.ToString('u'))"
     }
 
     # return an object that contains the 'expires_on' property along with the token
@@ -311,8 +311,12 @@ function Connect-AZTenant {
     )
 
     # just return if we already have a valid Bearer token
-    if ($script:AZToken -and (Get-Date) -lt $script:AZToken.Expires) {
-        return
+    if ($script:AZToken ) {
+        Write-Debug "Token Expires: $($script:AZToken.Expires)"
+        if ((Get-DateTimeOffsetNow) -lt $script:AZToken.Expires) {
+            Write-Debug "Existing token has not expired."
+            return
+        }
     }
 
     switch ($PSCmdlet.ParameterSetName) {
@@ -346,9 +350,11 @@ function Connect-AZTenant {
         }
     }
 
+    Write-Debug "Retrieved token expiration: $token.expires_on"
+
     # create a token object that we can use for subsequence calls with a 5 min buffer on the expiration
     $script:AZToken = [pscustomobject]@{
-        Expires = (Get-Date '1/1/1970').AddSeconds($token.expires_on - 300)
+        Expires    = [DateTimeOffset]::FromUnixTimeSeconds($token.expires_on).AddMinutes(-5)
         AuthHeader = @{ Authorization = "Bearer $($token.access_token)" }
     }
 }