+++ Posh-ACME/Posh-ACME.psm1
@@ -1,5 +1,28 @@
 #Requires -Version 5.1
 
+# Before we do anything else, make sure we have a sufficient .NET version that can load
+# the .NET Standard 2.0 version of Bouncy Castle we're using. It's supposed to be compatible
+# with .NET 4.6.1, but only if the app is compiled to support it (which PowerShell is not).
+# So it only loads properly on .NET 4.7.1 or later which is also the minimum version we need
+# to fully support ECC based certs. Any version of .NET Core should already work.
+if ($PSVersionTable.PSEdition -eq 'Desktop') {
+    # https://docs.microsoft.com/en-us/dotnet/framework/migration-guide/how-to-determine-which-versions-are-installed#to-check-for-a-minimum-required-net-framework-version-by-querying-the-registry-in-powershell-net-framework-45-and-later
+    $netBuild = (Get-ItemProperty "HKLM:SOFTWARE\Microsoft\NET Framework Setup\NDP\v4\Full").Release
+    if ($netBuild -ge 461308) { <# 4.7.1+ - all good #> }
+    else {
+        if     ($netBuild -ge 460798) { $netVer = '4.7' }
+        elseif ($netBuild -ge 394802) { $netVer = '4.6.2' }
+        elseif ($netBuild -ge 394254) { $netVer = '4.6.1' }
+        elseif ($netBuild -ge 393295) { $netVer = '4.6' }
+        elseif ($netBuild -ge 379893) { $netVer = '4.5.2' }
+        elseif ($netBuild -ge 378675) { $netVer = '4.5.1' }
+        Write-Warning "**********************************************************************"
+        Write-Warning "Insufficient .NET version. Found .NET $netVer (build $netBuild)."
+        Write-Warning ".NET 4.7.1 or later is required to ensure proper functionality."
+        Write-Warning "**********************************************************************"
+    }
+}
+
 # Get public and private function definition files.
 $Public  = @( Get-ChildItem -Path $PSScriptRoot\Public\*.ps1 -ErrorAction SilentlyContinue )
 $Private = @( Get-ChildItem -Path $PSScriptRoot\Private\*.ps1 -ErrorAction SilentlyContinue )
@@ -23,7 +46,7 @@ $script:WellKnownDirs = @{
     LE_STAGE = 'https://acme-staging-v02.api.letsencrypt.org/directory';
 }
 $script:HEADER_NONCE = 'Replay-Nonce'
-$script:USER_AGENT = "Posh-ACME/2.4.0 PowerShell/$($PSVersionTable.PSVersion)"
+$script:USER_AGENT = "Posh-ACME/2.7.0 PowerShell/$($PSVersionTable.PSVersion)"
 $script:COMMON_HEADERS = @{'Accept-Language'='en-us,en;q=0.5'}
 
 # Invoke-WebRequest and Invoke-RestMethod on PowerShell 5.1 both use