+++ Posh-ACME/Private/Get-DateTimeOffsetNow.ps1
@@ -1,0 +1,6 @@
+# This function only exists so that we can mock it in Pester tests
+function Get-DateTimeOffsetNow {
+    [CmdletBinding()]
+    param()
+    [System.DateTimeOffset]::Now
+}