+++ Posh-ACME/Private/Update-PAAccount.ps1
@@ -17,7 +17,8 @@ function Update-PAAccount {
         # grab the account from explicit parameters or the current memory copy
         if (!$ID) {
             if (!$script:Acct -or !$script:Acct.id) {
-                throw "No ACME account configured. Run Set-PAAccount or specify an ID."
+                Write-Warning "No ACME account configured. Run Set-PAAccount or specify an ID."
+                return
             }
             $acct = $script:Acct
             $UpdatingCurrent = $true
@@ -30,13 +31,15 @@ function Update-PAAccount {
             } else {
                 $UpdatingCurrent = $false
                 $acct = Get-PAAccount $ID
+                if ($null -eq $acct) {
+                    Write-Warning "Specified account id ($ID) not found. Nothing to update."
+                    return
+                }
+
             }
         }
 
         Write-Debug "Refreshing account $($acct.id)"
-
-        # hydrate the account key
-        $acctKey = $acct.key | ConvertFrom-Jwk
 
         # build the header
         $header = @{
@@ -51,7 +54,7 @@ function Update-PAAccount {
 
         # send the request
         try {
-            $response = Invoke-ACME $header.url $acctKey $header $payloadJson -EA Stop
+            $response = Invoke-ACME $header $payloadJson $acct -EA Stop
         } catch { throw }
         Write-Debug "Response: $($response.Content)"