+++ Posh-ACME/Public/Set-PAOrder.ps1
@@ -37,6 +37,11 @@ function Set-PAOrder {
             # order specified
             $order = Get-PAOrder $MainDomain
 
+            if ($null -eq $order) {
+                Write-Warning "Specified order for $MainDomain was not found. No changes made."
+                return
+            }
+
         } elseif (!$script:Order -or ($MainDomain -and ($MainDomain -ne $script:Order.MainDomain))) {
             # This is a definite order switch
 
@@ -103,9 +108,6 @@ function Set-PAOrder {
             }
             $certStr = $certLines[$certStart..$certEnd] -join '' | ConvertTo-Base64Url -FromBase64
 
-            # hydrate the account key
-            $acctKey = $acct.key | ConvertFrom-Jwk
-
             # build the header
             $header = @{
                 alg   = $acct.alg;
@@ -118,7 +120,7 @@ function Set-PAOrder {
 
             # send the request
             try {
-                $response = Invoke-ACME $header.url $acctKey $header $payloadJson -EA Stop
+                $response = Invoke-ACME $header $payloadJson $acct -EA Stop
             } catch { throw }
             Write-Debug "Response: $($response.Content)"